7zpp: don't truncate sizes larger than maxuint32
Archive members may be larger than 2**32-1. kpidSize is a VT_UI8 so per
MSDN, the correct field of the union is uhVal.

https://msdn.microsoft.com/en-us/library/windows/desktop/aa380072(v=vs.85).aspx

+++ src/7zip-cpp/7zpp/SevenZipArchive.cpp
@@ -64,7 +64,7 @@ namespace SevenZip
 		return m_itemnames;
 	}
 
-	std::vector<size_t> SevenZipArchive::GetOrigSizes()
+	std::vector<UINT64> SevenZipArchive::GetOrigSizes()
 	{
 		if (!m_ReadMetadata)
 		{
